requirements.txt
streamlit==1.32.0
qrcode==7.4.2
pillow==10.2.0 